pyproject.toml
[project]
name = "medfusion"
version = "0.1.0"
description = "A pluggable, modular framework for multimodal medical research with deep learning"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Medical AI Research", email = "iridite@foxmail.com" }
]
keywords = [
    "medical-ai",
    "deep-learning",
    "multimodal-learning",
    "medical-imaging",
    "research-framework",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "numpy>=1.24.0,<2.4",
    "pandas>=2.0.0,<3.0",
    "pillow>=10.0.0,<12.0",
    "scikit-learn>=1.3.0",
    "scikit-image>=0.21.0",
    "scipy>=1.11.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.13.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "opencv-python>=4.8.0",
    "tensorboard>=2.14.0",
    "pydicom>=3.0.1",
    "einops>=0.7.0",
    "timm>=0.9.0",
]

[project.optional-dependencies]
wandb = ["wandb>=0.15.0"]
lightning = ["pytorch-lightning>=2.0.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
med-train = "med_core.cli:train"
med-evaluate = "med_core.cli:evaluate"
med-preprocess = "med_core.cli:preprocess"



[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[[tool.uv.index]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008", "B905"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = ["--verbose", "--color=yes"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true
ignore_missing_imports = true

# Per-module options
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "torchvision.*",
    "monai.*",
    "tensorboard.*",
    "cv2.*",
    "sklearn.*",
    "scipy.*",
    "matplotlib.*",
    "seaborn.*",
    "PIL.*",
]
ignore_missing_imports = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
